+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGMDW3FQRMNV3PH68KS3JBW7"
+        "01KGP343BRN8H05PPADD5FPZ23"
     )
     r"""The unique identifier of the data source"""