+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGMDW3FN71QSH3FD9KSXZR9F"
+        "01KGP343BN73YPE7XN7BC3T670"
     )
     r"""The unique identifier of the data source"""