+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGMDW3FP00QEMN8EFF200J92"
+        "01KGP343BQY9HKP9ZS1ED1RCSA"
     )
     r"""The unique identifier of the data source"""