+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGMDW3FTZDJ5RFXMRJ2GSFQP"
+        "01KGP343BY7SN1SG84Y19FZ1BR"
     )
     r"""The unique identifier of the data source"""