+++ packages/orq-rc/RELEASES.md
@@ -5269,3 +5269,13 @@ Based on:
 - [python v4.3.0-rc.25] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.25 - packages/orq-rc
+
+## 2026-02-05 05:09:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.704.0 (2.806.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.26] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.26 - packages/orq-rc